Refs #74. Checkpointing progress on peak fitting test.
